requirements.txt
streamlit==1.32.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf==3.17.4
arxiv==2.0.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
openai==0.28.1
python-dotenv==1.0.0
tiktoken==0.5.2
crewai==0.28.0